+++ skbio/stats/distance/_base.py
@@ -21,6 +21,7 @@ from skbio.util import find_duplicates
 from skbio.util._decorator import experimental, classonlymethod
 from skbio.util._misc import resolve_key
 
+from ._utils import is_symmetric_and_hollow
 from ._utils import distmat_reorder, distmat_reorder_condensed
 
 
@@ -71,6 +72,9 @@ class DissimilarityMatrix(SkbioObject):
         rows/cols in `data`. If ``None`` (the default), IDs will be
         monotonically-increasing integers cast as strings, with numbering
         starting from zero, e.g., ``('0', '1', '2', '3', ...)``.
+    validate : bool, optional
+        If `validate` is ``True`` (the default) and data is not a
+        DissimilarityMatrix object, the input data will be validated.
 
     See Also
     --------
@@ -94,8 +98,21 @@ class DissimilarityMatrix(SkbioObject):
     _matrix_element_name = 'dissimilarity'
 
     @experimental(as_of="0.4.0")
-    def __init__(self, data, ids=None):
+    def __init__(self, data, ids=None, validate=True):
+        validate_full = validate
+        validate_shape = False
+        validate_ids = False
+
         if isinstance(data, DissimilarityMatrix):
+            if isinstance(data, self.__class__):
+                # Never validate when copying from an object
+                # of the same type
+                # We should be able to assume it is already
+                # in a good state.
+                validate_full = False
+                validate_shape = False
+                # but do validate ids, if redefining them
+                validate_ids = False if ids is None else True
             ids = data.ids if ids is None else ids
             data = data.data
 
@@ -121,12 +138,28 @@ class DissimilarityMatrix(SkbioObject):
             data = np.asarray(data, dtype='float')
 
         if data.ndim == 1:
+            # We can assume squareform will return a symmetric square matrix
+            # so no need for full validation.
+            # Still do basic checks (e.g. zero length)
+            # and id validation
             data = squareform(data, force='tomatrix', checks=False)
+            validate_full = False
+            validate_shape = True
+            validate_ids = True
+
         if ids is None:
             ids = (str(i) for i in range(data.shape[0]))
+            # I just created the ids, so no need to re-validate them
+            validate_ids = False
         ids = tuple(ids)
 
-        self._validate(data, ids)
+        if validate_full:
+            self._validate(data, ids)
+        else:
+            if validate_shape:
+                self._validate_shape(data)
+            if validate_ids:
+                self._validate_ids(data, ids)
 
         self._data = data
         self._ids = ids
@@ -216,7 +249,7 @@ class DissimilarityMatrix(SkbioObject):
     @ids.setter
     def ids(self, ids_):
         ids_ = tuple(ids_)
-        self._validate(self.data, ids_)
+        self._validate_ids(self.data, ids_)
         self._ids = ids_
         self._id_index = self._index_list(self._ids)
 
@@ -278,7 +311,10 @@ class DissimilarityMatrix(SkbioObject):
             `self`.
 
         """
-        return self.__class__(self.data.T.copy(), deepcopy(self.ids))
+        # Note: Skip validation, since we assume self was already validated
+        return self.__class__(self.data.T.copy(),
+                              deepcopy(self.ids),
+                              validate=False)
 
     @experimental(as_of="0.4.0")
     def index(self, lookup_id):
@@ -344,7 +380,10 @@ class DissimilarityMatrix(SkbioObject):
         """
         # We deepcopy IDs in case the tuple contains mutable objects at some
         # point in the future.
-        return self.__class__(self.data.copy(), deepcopy(self.ids))
+        # Note: Skip validation, since we assume self was already validated
+        return self.__class__(self.data.copy(),
+                              deepcopy(self.ids),
+                              validate=False)
 
     @experimental(as_of="0.4.0")
     def filter(self, ids, strict=True):
@@ -389,9 +428,11 @@ class DissimilarityMatrix(SkbioObject):
                     pass
             ids = found_ids
 
+        # Note: Skip validation, since we assume self was already validated
+        # But ids are new, so validate them explicitly
         filtered_data = distmat_reorder(self._data, idxs)
-
-        return self.__class__(filtered_data, ids)
+        self._validate_ids(filtered_data, ids)
+        return self.__class__(filtered_data, ids, validate=False)
 
     def _stable_order(self, ids):
         """Obtain a stable ID order with respect to self
@@ -859,15 +900,43 @@ class DissimilarityMatrix(SkbioObject):
         else:
             return self.data.__getitem__(index)
 
-    def _validate(self, data, ids):
-        """Validate the data array and IDs.
+    def _validate_ids(self, data, ids):
+        """Validate the IDs.
 
-        Checks that the data is at least 1x1 in size, 2D, square, hollow, and
-        contains only floats. Also checks that IDs are unique and that the
+        Checks that IDs are unique and that the
         number of IDs matches the number of rows/cols in the data array.
 
         Subclasses can override this method to perform different/more specific
-        validation (e.g., see `DistanceMatrix`).
+        validation.
+
+        Notes
+        -----
+        Accepts arguments instead of inspecting instance attributes to avoid
+        creating an invalid dissimilarity matrix before raising an error.
+        Otherwise, the invalid dissimilarity matrix could be used after the
+        exception is caught and handled.
+
+        """
+        duplicates = find_duplicates(ids)
+        if duplicates:
+            formatted_duplicates = ', '.join(repr(e) for e in duplicates)
+            raise DissimilarityMatrixError("IDs must be unique. Found the "
+                                           "following duplicate IDs: %s" %
+                                           formatted_duplicates)
+        if 0 == len(ids):
+            raise DissimilarityMatrixError("IDs must be at least 1 in "
+                                           "size.")
+        if len(ids) != data.shape[0]:
+            raise DissimilarityMatrixError("The number of IDs (%d) must match "
+                                           "the number of rows/columns in the "
+                                           "data (%d)." %
+                                           (len(ids), data.shape[0]))
+
+    def _validate_shape(self, data):
+        """Validate the data array shape.
+
+        Checks that the data is at least 1x1 in size, 2D, square, and
+        contains only floats.
 
         Notes
         -----
@@ -890,17 +959,27 @@ class DissimilarityMatrix(SkbioObject):
         if data.dtype not in (np.float32, np.float64):
             raise DissimilarityMatrixError("Data must contain only floating "
                                            "point values.")
-        duplicates = find_duplicates(ids)
-        if duplicates:
-            formatted_duplicates = ', '.join(repr(e) for e in duplicates)
-            raise DissimilarityMatrixError("IDs must be unique. Found the "
-                                           "following duplicate IDs: %s" %
-                                           formatted_duplicates)
-        if len(ids) != data.shape[0]:
-            raise DissimilarityMatrixError("The number of IDs (%d) must match "
-                                           "the number of rows/columns in the "
-                                           "data (%d)." %
-                                           (len(ids), data.shape[0]))
+
+    def _validate(self, data, ids):
+        """Validate the data array and IDs.
+
+        Checks that the data is at least 1x1 in size, 2D, square, and
+        contains only floats. Also checks that IDs are unique and that the
+        number of IDs matches the number of rows/cols in the data array.
+
+        Subclasses can override this method to perform different/more specific
+        validation (e.g., see `DistanceMatrix`).
+
+        Notes
+        -----
+        Accepts arguments instead of inspecting instance attributes to avoid
+        creating an invalid dissimilarity matrix before raising an error.
+        Otherwise, the invalid dissimilarity matrix could be used after the
+        exception is caught and handled.
+
+        """
+        self._validate_shape(data)
+        self._validate_ids(data, ids)
 
     def _index_list(self, list_):
         return {id_: idx for idx, id_ in enumerate(list_)}
@@ -1071,8 +1150,9 @@ class DistanceMatrix(DissimilarityMatrix):
             permuted_condensed = distmat_reorder_condensed(self._data, order)
             return permuted_condensed
         else:
+            # Note: Skip validation, since we assume self was already validated
             permuted = distmat_reorder(self._data, order)
-            return self.__class__(permuted, self.ids)
+            return self.__class__(permuted, self.ids, validate=False)
 
     def _validate(self, data, ids):
         """Validate the data array and IDs.
@@ -1083,11 +1163,13 @@ class DistanceMatrix(DissimilarityMatrix):
         """
         super(DistanceMatrix, self)._validate(data, ids)
 
-        if (data.T != data).any():
+        data_sym, data_hol = is_symmetric_and_hollow(data)
+
+        if not data_sym:
             raise DistanceMatrixError(
                 "Data must be symmetric and cannot contain NaNs.")
 
-        if np.trace(data) != 0:
+        if not data_hol:
             raise DistanceMatrixError("Data must be hollow (i.e., the diagonal"
                                       " can only contain zeros).")