+++ skbio/stats/distance/tests/test_base.py
@@ -24,6 +24,7 @@ from skbio.stats.distance import (
     DissimilarityMatrix, randdm)
 from skbio.stats.distance._base import (_preprocess_input,
                                         _run_monte_carlo_stats)
+from skbio.stats.distance._utils import is_symmetric_and_hollow
 from skbio.util import assert_data_frame_almost_equal
 from skbio.util._testing import assert_series_almost_equal
 
@@ -228,6 +229,8 @@ class DissimilarityMatrixTests(DissimilarityMatrixTestData):
         data = [[1, 1], [1, 1]]
         obs = DissimilarityMatrix(data, ['a', 'b'])
         self.assertTrue(np.array_equal(obs.data, data))
+        data_hollow = skbio.stats.distance._utils.is_hollow(obs.data)
+        self.assertEqual(data_hollow, False)
 
     def test_init_no_ids(self):
         exp = DissimilarityMatrix(self.dm_3x3_data, ('0', '1', '2'))
@@ -717,6 +720,34 @@ class DissimilarityMatrixTests(DissimilarityMatrixTestData):
         with self.assertRaises(DissimilarityMatrixError):
             self.dm_3x3._validate(np.array([[0, 42], [42, 0]]), ['a', 'b'])
 
+    def test_validate_invalid_shape(self):
+        # first check it actually likes good matrices
+        self.dm_3x3._validate_shape(np.array([[0., 42.], [42., 0.]]))
+        # it checks just the shape, not the content
+        self.dm_3x3._validate_shape(np.array([[1., 2.], [3., 4.]]))
+        # empty array
+        with self.assertRaises(DissimilarityMatrixError):
+            self.dm_3x3._validate_shape(np.array([]))
+        # invalid shape
+        with self.assertRaises(DissimilarityMatrixError):
+            self.dm_3x3._validate_shape(np.array([[0., 42.],
+                                                  [42., 0.],
+                                                  [22., 22.]]))
+        with self.assertRaises(DissimilarityMatrixError):
+            self.dm_3x3._validate_shape(np.array([[[0., 42.], [42., 0.]],
+                                                  [[0., 24.], [24., 0.]]]))
+
+    def test_validate_invalid_ids(self):
+        # repeated ids
+        with self.assertRaises(DissimilarityMatrixError):
+            self.dm_3x3._validate_ids(self.dm_3x3.data, ['a', 'a'])
+        # empty ids
+        with self.assertRaises(DissimilarityMatrixError):
+            self.dm_3x3._validate_ids(self.dm_3x3.data, [])
+        # invalid shape
+        with self.assertRaises(DissimilarityMatrixError):
+            self.dm_3x3._validate_ids(self.dm_3x3.data, ['a', 'b', 'c', 'd'])
+
 
 class DistanceMatrixTests(DissimilarityMatrixTestData):
     def setUp(self):
@@ -978,6 +1009,55 @@ class DistanceMatrixTests(DissimilarityMatrixTestData):
 
         exp = pd.Series([0.123], index=pd.Index([('0', '1')]))
         assert_series_almost_equal(series, exp)
+
+    def test_validate_asym_shape(self):
+        # first check it actually likes good matrices
+        data_good = np.array([[0., 42.], [42., 0.]])
+        data_sym, data_hollow = is_symmetric_and_hollow(data_good)
+        self.assertEqual(data_sym, True)
+        del data_sym
+        self.assertEqual(data_hollow, True)
+        del data_hollow
+        data_sym = skbio.stats.distance._utils.is_symmetric(data_good)
+        self.assertEqual(data_sym, True)
+        del data_sym
+        data_hollow = skbio.stats.distance._utils.is_hollow(data_good)
+        self.assertEqual(data_hollow, True)
+        del data_hollow
+        self.dm_3x3._validate_shape(data_good)
+        del data_good
+
+        # _validate_shap checks just the shape, not the content
+        bad_data = np.array([[1., 2.], [3., 4.]])
+        data_sym, data_hollow = is_symmetric_and_hollow(bad_data)
+        self.assertEqual(data_sym, False)
+        del data_sym
+        self.assertEqual(data_hollow, False)
+        del data_hollow
+        data_sym = skbio.stats.distance._utils.is_symmetric(bad_data)
+        self.assertEqual(data_sym, False)
+        del data_sym
+        data_hollow = skbio.stats.distance._utils.is_hollow(bad_data)
+        self.assertEqual(data_hollow, False)
+        del data_hollow
+        self.dm_3x3._validate_shape(bad_data)
+        del bad_data
+
+        # re-try with partially bad data
+        bad_data = np.array([[0., 2.], [3., 0.]])
+        data_sym, data_hollow = is_symmetric_and_hollow(bad_data)
+        self.assertEqual(data_sym, False)
+        del data_sym
+        self.assertEqual(data_hollow, True)
+        del data_hollow
+        data_sym = skbio.stats.distance._utils.is_symmetric(bad_data)
+        self.assertEqual(data_sym, False)
+        del data_sym
+        data_hollow = skbio.stats.distance._utils.is_hollow(bad_data)
+        self.assertEqual(data_hollow, True)
+        del data_hollow
+        self.dm_3x3._validate_shape(bad_data)
+        del bad_data
 
 
 class RandomDistanceMatrixTests(TestCase):